+++ packages/marionette_cli/test/cli/marionette_command_runner_test.dart
@@ -29,7 +29,6 @@ void main() {
         'get-interactive-elements',
         'tap',
         'secondary-tap',
-        'tertiary-tap',
         'double-tap',
         'long-press',
         'pinch-zoom',